fix repeat, mypy base
